==> notebooks/03_model_fitting_and_analysis/04_gather_output.ipynb <==
# Gather Output

This notebook gathers all figures and tables used in the study.


# Define function
import sys
import shutil

sys.path.insert(0, "../../src")
from imports import *

init_notebook()


def copy_and_rename_file(source: str, target: str) -> None:
    """
    Copies a plot file from the source path to the target path.

    Parameters:
        source (str): Path to the original plot file.
        target (str): Path where the copied file should be saved, including the new name if renaming.
    """
    display("----------------------------------------------------------------")
    # Report if file missing
    if not os.path.isfile(source):
        print(f"🔴 Error: Source file '{source}' not found.")
        return

    elif os.path.isfile(target):
        shutil.copy(source, target)
        print(
            f"🟡 Warning: Target file '{target}' was overwritten.\n - From '{source}'\n - To {target}"
        )
    else:
        shutil.copy(source, target)
        print(f"🟢 File copied successfully!\n - From '{source}'\n - To {target}")

# Create output folder
os.makedirs("../../output", exist_ok=True)

# Get latest model_analysis folder
latest_model_analysis = glob.glob(f"./model_analysis/*")[-1]
base_dir = "./model_analysis/*/*/*/*/*"

# Copy the plots from the latest model_analysis folder to the plots folder
# ! MAIN ---
# FIGURE 1 (created in notebook `02_calculate_mortality.ipynb`)
copy_and_rename_file(
    "../01_process_nfi_data/overview-climate_change-mortality.png",
    "../../output/fig-1.png",
)

# FIGURE 2 (created in notebook `02_random_forest_analysis.ipynb`)
copy_and_rename_file(
    sorted(glob.glob(f"{base_dir}/top9.png"))[-1],
    "../../output/fig-2.png",
)

# FIGURE 3 (created in notebook `02_random_forest_analysis.ipynb`)
copy_and_rename_file(
    sorted(
        glob.glob(
            f"{base_dir}/pie-temp-double_change-top9_True-scaled_by_100perc_True-scale_by_nmodels_False-add_boxes_False.png"
        )
    )[-1],
    "../../output/fig-3.png",
)

# FIGURE 4 (created in notebook `02_random_forest_analysis.ipynb`)
copy_and_rename_file(
    sorted(
        glob.glob(
            f"{base_dir}/pie-spei-double_change-top9_True-scaled_by_100perc_True-scale_by_nmodels_False-add_boxes_False.png"
        )
    )[-1],
    "../../output/fig-4.png",
)

# FIGURE 5 (created in notebook `02_random_forest_analysis.ipynb`)
copy_and_rename_file(
    sorted(glob.glob(f"{base_dir}/wetter-labels_top9+_long_True-colored_mixed.png"))[
        -1
    ],
    "../../output/fig-5.png",
)

# FIGURE 6 (created in notebook `02_random_forest_analysis.ipynb`)
copy_and_rename_file(
    sorted(glob.glob(f"{base_dir}/top9-corr_r-wetter_vs_drier-subset_of_runs_all.png"))[
        -1
    ],
    "../../output/fig-6.png",
)

# ! SUPPLEMENTARY ---
# FIGURES S1 (created in notebook `02_calculate_mortality.ipynb`)
copy_and_rename_file(
    "../01_process_nfi_data/mortality_trends-all_species-1.png",
    "../../output/fig-s1-1.png",
)
copy_and_rename_file(
    "../01_process_nfi_data/mortality_trends-all_species-2.png",
    "../../output/fig-s1-2.png",
)

# FIGURE S2 (created in notebook `02_calculate_mortality.ipynb`)
copy_and_rename_file(
    "../01_process_nfi_data/regional_mortality_trends.png",
    "../../output/fig-s2.png",
)

# FIGURES S3 (created in notebook `02_calculate_mortality.ipynb`)
copy_and_rename_file(
    "../01_process_nfi_data/trends_harvest_comparison-1.png",
    "../../output/fig-s3-1.png",
)

copy_and_rename_file(
    "../01_process_nfi_data/trends_harvest_comparison-2.png",
    "../../output/fig-s3-2.png",
)

# FIGURES S4 (created in notebook `02_calculate_mortality.ipynb`)
copy_and_rename_file(
    "../01_process_nfi_data/trends_by_height_class-1.png",
    "../../output/fig-s4-1.png",
)

copy_and_rename_file(
    "../01_process_nfi_data/trends_by_height_class-2.png",
    "../../output/fig-s4-2.png",
)

# FIGURES S5 (created in notebook `02_calculate_mortality.ipynb`)
copy_and_rename_file(
    "../01_process_nfi_data/trends_ba_absolute-all_species-1.png",
    "../../output/fig-s5-1.png",
)

copy_and_rename_file(
    "../01_process_nfi_data/trends_ba_absolute-all_species-2.png",
    "../../output/fig-s5-2.png",
)

# FIGURE S6 (created in notebook `02_random_forest_analysis.ipynb`)
copy_and_rename_file(
    sorted(glob.glob(f"{base_dir}/all.png"))[-1],
    "../../output/fig-s6.png",
)

# FIGURE S7 (created in notebook `02_random_forest_analysis.ipynb`)
copy_and_rename_file(
    sorted(glob.glob(f"{base_dir}/feature_counts.png"))[-1],
    "../../output/fig-s7.png",
)

# FIGURE S8 (created in notebook `02_random_forest_analysis.ipynb`)
copy_and_rename_file(
    sorted(
        glob.glob(
            f"{base_dir}/pie-spei-double_change-top9_False-scaled_by_100perc_True-scale_by_nmodels_False-add_boxes_False.png"
        )
    )[-1],
    "../../output/fig-s8.png",
)

# FIGURE S9 (created in notebook `02_random_forest_analysis.ipynb`)
copy_and_rename_file(
    sorted(
        glob.glob(
            f"{base_dir}/pie-temp-double_change-top9_False-scaled_by_100perc_True-scale_by_nmodels_False-add_boxes_False.png"
        )
    )[-1],
    "../../output/fig-s9.png",
)

# FIGURE S10 (created in notebook `02_random_forest_analysis.ipynb`)
copy_and_rename_file(
    sorted(glob.glob(f"{base_dir}/temporal-temp-all.png"))[-1],
    "../../output/fig-s10.png",
)

# FIGURE S11 (created in notebook `02_random_forest_analysis.ipynb`)
copy_and_rename_file(
    sorted(glob.glob(f"{base_dir}/temporal-spei-warmer_drier.png"))[-1],
    "../../output/fig-s11.png",
)

# FIGURE S12 (created in notebook `02_random_forest_analysis.ipynb`)
copy_and_rename_file(
    sorted(glob.glob(f"{base_dir}/temporal-spei-warmer_wetter.png"))[-1],
    "../../output/fig-s12.png",
)

# FIGURE S13 (created in notebook `02_random_forest_analysis.ipynb`)
copy_and_rename_file(
    sorted(glob.glob(f"{base_dir}/all-corr_r-wetter_vs_drier-subset_of_runs_all.png"))[
        -1
    ],
    "../../output/fig-s13.png",
)

# Figure S14 (created in notebook `10_get_spei.ipynb`)
copy_and_rename_file(
    "./../02_collect_features/climate_evolution_figures/S_spei_timeseries.png",
    "../../output/fig-s14.png",
)

# FIGURE S15 (conceptual figure by hand)

# FIGURE S16 (created in notebook `01_clean_raw_nfi_data.ipynb`)
copy_and_rename_file(
    "./../01_process_nfi_data/rf_height_prediction.png",
    "../../output/fig-s16.png",
)

# FIGURE S17 (created in notebook `10_get_climate.ipynb`)
copy_and_rename_file(
    "./../02_collect_features/example_temperature_features.png",
    "../../output/fig-s17.png",
)

# FIGURE S18 (created in notebook `10_get_spei.ipynb`)
copy_and_rename_file(
    "./../02_collect_features/example_spei_features.png",
    "../../output/fig-S18.png",
)

# FIGURE S19 (conceptual figure by hand)

# Figure S20 (created in notebook `02_random_forest_analysis.ipynb`)
copy_and_rename_file(
    sorted(glob.glob(f"./model_analysis/*/*/*/*/smotek_md_distr_pca.png"))[-1],
    "../../output/fig-s20.png",
)

# Figure S21 (created in notebook `02_random_forest_analysis.ipynb`)
copy_and_rename_file(
    sorted(glob.glob(f"./model_analysis/*/*/*/*/smotek_md_boxplots.png"))[-1],
    "../../output/fig-s21.png",
)

# Figure S22 (created in notebook `02_random_forest_analysis.ipynb`)
copy_and_rename_file(
    sorted(glob.glob(f"./model_analysis/*/*/*/*/model_performance_boxplots.png"))[-1],
    "../../output/fig-s22.png",
)

# Figure S22 (created in notebook `02_random_forest_analysis.ipynb`)
copy_and_rename_file(
    sorted(glob.glob(f"./model_analysis/*/*/*/*/model_performance_boxplots.png"))[-1],
    "../../output/fig-s22.png",
)

# Figure S23 (created in notebook `02_random_forest_analysis.ipynb`)
copy_and_rename_file(
    sorted(
        glob.glob(
            f"./extra_runs/rfe_nkeep_4/_analysis/_top9_species/*/pattern_analysis/by_mk/roc_0.6-min_group_share_0.6/si_summary.png"
        )
    )[-1],
    "../../output/fig-s23.png",
)

# Figure S24 (created in notebook `02_random_forest_analysis.ipynb`)
copy_and_rename_file(
    sorted(
        glob.glob(
            f"./extra_runs/rfe_nkeep_6/_analysis/_top9_species/*/pattern_analysis/by_mk/roc_0.6-min_group_share_0.6/si_summary.png"
        )
    )[-1],
    "../../output/fig-s24.png",
)

# Figure S25 (created in notebook `02_random_forest_analysis.ipynb`)
copy_and_rename_file(
    sorted(glob.glob(f"./model_analysis/*/*/*/*/model_performance_rfe_n.png"))[-1],
    "../../output/fig-s25.png",
)

# Figure S26 (created in notebook `02_random_forest_analysis.ipynb`)
copy_and_rename_file(
    sorted(
        glob.glob(
            f"./extra_runs/smotek_1/_analysis/_top9_species/*/pattern_analysis/by_mk/roc_0.6-min_group_share_0.6/si_summary.png"
        )
    )[-1],
    "../../output/fig-s26.png",
)

# Figure S27 (created in notebook `02_random_forest_analysis.ipynb`)
copy_and_rename_file(
    sorted(
        glob.glob(
            f"./extra_runs/smotek_15/_analysis/_top9_species/*/pattern_analysis/by_mk/roc_0.6-min_group_share_0.6/si_summary.png"
        )
    )[-1],
    "../../output/fig-s27.png",
)

# Figure S28 (created in notebook `02_random_forest_analysis.ipynb`)
copy_and_rename_file(
    sorted(
        glob.glob(
            f"./extra_runs/smotek_0/_analysis/_top9_species/*/pattern_analysis/by_mk/roc_0.6-min_group_share_0.6/si_summary.png"
        )
    )[-1],
    "../../output/fig-s28.png",
)

# Figure s29 (created in notebook `02_random_forest_analysis.ipynb`)
copy_and_rename_file(
    sorted(glob.glob(f"./model_analysis/*/*/*/*/model_performance_smote_k.png"))[-1],
    "../../output/fig-s29.png",
)

# FIGURE S30 (created in notebook `03_glmm_analysis.ipynb`)
copy_and_rename_file(
    sorted(
        glob.glob(
            f"./model_analysis/*/glmm_analysis/glmm_rf_model_performance_comparison.png"
        )
    )[-1],
    "../../output/fig-s30.png",
)

# FIGURE S31 (created in notebook `03_glmm_analysis.ipynb`)
copy_and_rename_file(
    sorted(glob.glob(f"./model_analysis/*/glmm_analysis/si_summary_importance.png"))[
        -1
    ],
    "../../output/fig-s31.png",
)

## Tables


# TABLE S1 (created in notebook `02_random_forest_analysis.ipynb`)
copy_and_rename_file(
    glob.glob(f"{base_dir}/mortality_trends.csv")[0],
    "../../output/table-s1.csv",
)

# TABLE S2 (written by hand)

# TABLE S3 (created in notebook `02_random_forest_analysis.ipynb`)
copy_and_rename_file(
    glob.glob(f"{base_dir}/category_importance.csv")[0],
    "../../output/table-s3.csv",
)

# TABLE S4 (created in notebook `02_random_forest_analysis.ipynb`)
copy_and_rename_file(
    glob.glob(f"{base_dir}/model_performance_summary.csv")[0],
    "../../output/table-s4.csv",
)

# TABLE S5 (created by hand)

# TABLE S6 (created in notebook `02_random_forest_analysis.ipynb`)
copy_and_rename_file(
    glob.glob(f"{base_dir}/spei_features_all_species.csv")[0],
    "../../output/table-s6.csv",
)

# TABLE S7 (created in notebook `02_random_forest_analysis.ipynb`)
copy_and_rename_file(
    glob.glob(f"{base_dir}/temp_features_all_species.csv")[0],
    "../../output/table-s7.csv",
)

# TABLE S8 (created in notebook `02_random_forest_analysis.ipynb`)
copy_and_rename_file(
    glob.glob(f"{base_dir}/climate_features_all_species.csv")[0],
    "../../output/table-s8.csv",
)

# TABLE S9 (created in notebook `02_random_forest_analysis.ipynb`)
copy_and_rename_file(
    glob.glob(f"{base_dir}/linear_regression_results_height_wetter.csv")[0],
    "../../output/table-s9.csv",
)

---
